% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdownPkgTest-package.R
\docType{package}
\name{bookdownPkgTest-package}
\alias{bookdownPkgTest}
\alias{bookdownPkgTest-package}
\title{bookdownPkgTest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Francesca Vitalini \email{francesca.vitalini@mirai-solutions.com}

}
\keyword{internal}
